Implemented with BigDecimal Set MAX_PRECISION of division to 8 to fit on a nexus 4 screen Created a new Number class constructor that takes in a String
Implement numbers with BigDecimal #53
Implement Scientific notation numbers #28
The converted value should not be a double when not applicable #39

+++ app/src/main/java/com/samsmlee/unitcalculator/Unit/Unit.java
@@ -7,6 +7,8 @@ import com.samsmlee.unitcalculator.Converter.VolumeConverter;
 import com.samsmlee.unitcalculator.Unit.UnitBase.UnitBase;
 import com.samsmlee.unitcalculator.Unit.UnitType.UnitType;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Sam on 12/22/2014.
  */
@@ -51,7 +53,7 @@ public class Unit{
     }
 
 
-    public static double convert(Unit fromUnit, Unit toUnit, double value) {
+    public static BigDecimal convert(Unit fromUnit, Unit toUnit, BigDecimal value) {
 
         if (fromUnit == null) {
             throw new IllegalArgumentException("Cannot convert from a null unit");
@@ -63,7 +65,7 @@ public class Unit{
         if (!toUnit.equalType(fromUnit))
             throw new IllegalArgumentException("Incompatible type");
 
-        double convertedValue;
+        BigDecimal convertedValue;
         switch (fromUnit.getType()) {
 
             case LENGTH:
